+++ dev/webdock/domain/web-client-compose/hooks/test.sh
@@ -2,10 +2,7 @@
 set -e
 
 echo '[Hook] Running GUI tests...'
-sleep 15 # Selenium hack
 docker-compose exec web_tester \
-  npm run target-gui-tests-with-chrome
-docker-compose exec web_tester \
-  npm run target-gui-tests-with-firefox
+  npm run target-gui-tests
 
 echo '[Hook] Done.'